+++ src/containerapp/azext_containerapp/_constants.py
@@ -42,3 +42,4 @@ steps:
 DEFAULT_PORT = 8080  # used for no dockerfile scenario; not the hello world image
 
 HELLO_WORLD_IMAGE = "mcr.microsoft.com/azuredocs/containerapps-helloworld:latest"
+