+++ src/containerapp/azext_containerapp/_utils.py
@@ -1070,6 +1070,16 @@ def safe_get(model, *keys, default=None):
     return model.get(keys[-1], default)
 
 
+def safe_set(model, *keys, value):
+    penult = {}
+    for k in keys:
+        if k not in model:
+            model[k] = {}
+        penult = model
+        model = model[k]
+    penult[keys[-1]] = value
+
+
 def is_platform_windows():
     return platform.system() == "Windows"